fix: split oversized backend files for backend-size CI gate
Move session CLI wiring, session read projections, JavaScript world replay, and contract fixture helpers into dedicated files so backend-size limits pass on the PR head.
