+++ pkg/api/contracttests/openapi_contract_common_test.go
@@ -2,7 +2,6 @@ package apicontract_test
 
 import (
 	"context"
-	"encoding/json"
 	"os"
 	"sort"
 	"strings"
@@ -979,22 +978,4 @@ func assertJSONKeysPresent(t *testing.T, object map[string]any, name string, key
 			t.Fatalf("%s.%s is missing", name, key)
 		}
 	}
-}
-
-func loadCanonicalFactoryEventVocabularyFixture(t *testing.T) []map[string]any {
-	t.Helper()
-	fixtureBytes, err := os.ReadFile("../testdata/canonical-event-vocabulary-stream.json")
-	if err != nil {
-		t.Fatalf("read canonical event vocabulary fixture: %v", err)
-	}
-	assertJSONStringLiteralMissing(t, string(fixtureBytes), retiredFactoryEventTypeValues...)
-
-	var fixture []map[string]any
-	if err := json.Unmarshal(fixtureBytes, &fixture); err != nil {
-		t.Fatalf("parse canonical event vocabulary fixture: %v", err)
-	}
-	if len(fixture) != len(canonicalFactoryEventTypeValues) {
-		t.Fatalf("canonical event vocabulary fixture length = %d, want %d", len(fixture), len(canonicalFactoryEventTypeValues))
-	}
-	return fixture
 }
